requirements.txt
pip~=21.3
cryptography~=35.0.0
MarkupSafe~=2.0.1
Werkzeug~=2.0.2
Jinja2~=3.0.2
click~=8.0.3
itsdangerous~=2.0.1
greenlet~=1.1.2
setuptools~=57.4.0
Flask~=2.0.2
WTForms~=2.3.3
SQLAlchemy~=1.4.25
pyotp~=2.6.0
flask-login
pycryptodome